# doc/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandAck": {
        "description": "The reply to every command POST.\n\nDeliberately NOT the resulting state. That arrives on whichever socket the\npage has open, on the server's own schedule, so state has exactly one path\nand there is no ordering for a client to reconcile between two of them. What\nthis carries is only \"the command was understood and applied, and here is\nwhat it was\" -- enough to log and to assert on, and nothing a page renders.",
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "applied"
        ],
        "title": "CommandAck",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SeekBody": {
        "properties": {
          "offset_s": {
            "description": "Seconds from the start of the recording.",
            "minimum": 0.0,
            "title": "Offset S",
            "type": "number"
          }
        },
        "required": [
          "offset_s"
        ],
        "title": "SeekBody",
        "type": "object"
      },
      "SpeedBody": {
        "properties": {
          "speed": {
            "description": "Replay speed multiplier; 1 is real time.",
            "exclusiveMinimum": 0.0,
            "maximum": 10.0,
            "title": "Speed",
            "type": "number"
          }
        },
        "required": [
          "speed"
        ],
        "title": "SpeedBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
...
    "title": "p-SWAMP api",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/pmu-test-streamer/playback/back": {
      "post": {
        "description": "Step one frame back: the player reopens its stream one interval earlier.",
        "operationId": "pmu_test_streamer_back",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The command does not apply in the player's current mode."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Back",
        "tags": [
          "pmu-test-streamer"
        ]
      }
    },
    "/api/pmu-test-streamer/playback/forward": {
      "post": {
        "description": "Play one frame, independently of the play/pause state.",
        "operationId": "pmu_test_streamer_forward",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The command does not apply in the player's current mode."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forward",
        "tags": [
          "pmu-test-streamer"
        ]
      }
    },
...
    "/api/pmu-test-streamer/playback/play": {
      "post": {
        "description": "Start (or resume) this client's replay.",
        "operationId": "pmu_test_streamer_play",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The command does not apply in the player's current mode."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Play",
        "tags": [
          "pmu-test-streamer"
        ]
      }
    },
    "/api/pmu-test-streamer/playback/replay": {
      "post": {
        "description": "Switch this client back to the recording, paused at its start. 409 when\nno history source is configured.",
        "operationId": "pmu_test_streamer_replay",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The command does not apply in the player's current mode."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay",
        "tags": [
          "pmu-test-streamer"
        ]
      }
    },
    "/api/pmu-test-streamer/playback/seek": {
      "post": {
        "description": "Jump the replay to an offset into the recording. 409 while live.",
        "operationId": "pmu_test_streamer_seek",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeekBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The command does not apply in the player's current mode."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seek",
        "tags": [
          "pmu-test-streamer"
        ]
      }
    },
    "/api/pmu-test-streamer/playback/speed": {
      "post": {
        "description": "Change the replay speed. 409 while live.",
        "operationId": "pmu_test_streamer_speed",
        "parameters": [
          {
            "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
            "in": "query",
            "name": "client_id",
            "required": true,
            "schema": {
              "description": "The browser's client id -- the same value its WebSockets send, which is what makes a command apply to the pipeline the page is watching.",
              "pattern": "^\\d{1,20}$",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeedBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The command does not apply in the player's current mode."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Speed",
        "tags": [
          "pmu-test-streamer"
        ]
      }
    },
...
  },
...
}